chore: remove the unused code
