+++ frontend/appflowy_flutter/integration_test/hotkeys_test.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/workspace/application/settings/prelude.dart';
-import 'package:appflowy/workspace/presentation/home/menu/menu.dart';
+import 'package:appflowy/workspace/presentation/home/menu/sidebar/sidebar.dart';
 import 'package:appflowy/workspace/presentation/settings/settings_dialog.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/services.dart';
@@ -75,7 +75,7 @@ void main() {
 
       await tester.pumpAndSettle();
 
-      expect(find.byType(HomeMenu), findsOneWidget);
+      expect(find.byType(HomeSideBar), findsOneWidget);
 
       await FlowyTestKeyboard.simulateKeyDownEvent(
         [
@@ -89,7 +89,7 @@ void main() {
 
       await tester.pumpAndSettle();
 
-      expect(find.byType(HomeMenu), findsNothing);
+      expect(find.byType(HomeSideBar), findsNothing);
     });
   });
 }
